Fix Geckodriver download on Mac with Laravel Dusk 6.12.0+
In v6.12.0 of package "laravel/dusk", it changed the
return values public method `Laravel\Dusk\OperatingSystem::id()`
when running in a Mac environment. It will now _sometimes_
include the architecture in the string.

This caused this package's `dusk:firefox-driver` command
to skip the operating system configuration.

Remove the dependency on `OperatingSystem::id()` as
Geckodriver doesn't (yet) have a distributed ARM binary.