Fix always capturing output from 'tar' extraction
When decompressing the Geckodriver .tar.gz to
the local filesystem, suppress stderr output to
the console. Instead capture it for delayed
output in the thrown RuntimeException.

This indirectly also seems to fix extracting
binaries on Mac and Windows (native). The string
"x geckodriver" is being printed to the Mac Terminal
& Windows Command Prompt when PHP's exec('tar -v...')
is called. But the $output array is empty despite
the tarball including a file and tar's verbose is on.

From (probably untrustworthy) php.net user posts, it
may be caused by a newline character being printed on
the terminal before the handled filename "x geckodriver"
is output.

Unrelated: call Laravel's ProcessUtils::escapeArgument()
instead of PHP standard library's escapeshellarg()
because apparently the latter has buggy edge cases.